chore: remove redundant old implementation (#3802)
* chore: remove redundant old implementation

* chore: remove test cases related to pinterest_tag

* chore: fix test cases

* chore: fix lint issue

* chore: fix lint issue again
