--- src/v0/destinations/pinterest_tag/data/pinterestUserConfig.json
@@ -1,93 +1,0 @@
-[
-  {
-    "destKey": "em",
-    "sourceKeys": "emailOnly",
-    "sourceFromGenericMap": true,
-    "required": false
-  },
-  {
-    "destKey": "ph",
-    "sourceKeys": "phone",
-    "sourceFromGenericMap": true,
-    "required": false
-  },
-  {
-    "destKey": "ge",
-    "sourceKeys": ["traits.gender", "context.traits.gender"],
-    "required": false,
-    "metadata": {
-      "multikeyMap": [
-        {
-          "sourceVal": ["f", "F", "Female", "female", "FEMALE"],
-          "destVal": "f"
-        },
-        {
-          "sourceVal": ["m", "M", "Male", "male", "MALE"],
-          "destVal": "m"
-        },
-        {
-          "sourceVal": ["n", "N", "non-binary", "non binary", "Non-Binary", "Non Binary"],
-          "destVal": "n"
-        }
-      ]
-    }
-  },
-  {
-    "destKey": "db",
-    "sourceKeys": "birthday",
-    "required": false,
-    "sourceFromGenericMap": true
-  },
-  {
-    "destKey": "ln",
-    "sourceKeys": "lastName",
-    "required": false,
-    "sourceFromGenericMap": true
-  },
-  {
-    "destKey": "fn",
-    "sourceKeys": "firstName",
-    "required": false,
-    "sourceFromGenericMap": true
-  },
-  {
-    "destKey": "ct",
-    "sourceKeys": ["traits.address.city", "context.traits.address.city"],
-    "required": false
-  },
-  {
-    "destKey": "st",
-    "sourceKeys": ["traits.address.state", "context.traits.address.state"],
-    "required": false
-  },
-  {
-    "destKey": "zp",
-    "sourceKeys": ["traits.address.zip", "context.traits.address.zip"],
-    "required": false
-  },
-  {
-    "destKey": "country",
-    "sourceKeys": ["traits.address.country", "context.traits.address.country"],
-    "required": false
-  },
-  {
-    "destKey": "hashed_maids",
-    "sourceKeys": ["context.device.advertisingId"],
-    "required": false
-  },
-  {
-    "destKey": "client_ip_address",
-    "sourceKeys": ["context.ip", "request_ip"],
-    "required": false
-  },
-  {
-    "destKey": "client_user_agent",
-    "sourceKeys": "context.userAgent",
-    "required": false
-  },
-  {
-    "destKey": "external_id",
-    "sourceKeys": "userId",
-    "sourceFromGenericMap": true
-  }
-]
