+++ test/integrations/destinations/pinterest_tag/step/data.ts
@@ -75,6 +75,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -90,6 +91,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -101,6 +103,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -140,10 +145,16 @@ export const data = [
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
                       {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
                         quantity: 1,
                         item_price: '19',
                       },
                       {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
                         quantity: 2,
                         item_price: '3',
                       },
@@ -238,6 +249,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -253,6 +265,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -264,6 +277,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -301,10 +317,16 @@ export const data = [
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
                       {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
                         quantity: 1,
                         item_price: '19',
                       },
                       {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
                         quantity: 2,
                         item_price: '3',
                       },
@@ -380,6 +402,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -395,6 +418,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -406,6 +430,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -531,6 +558,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -546,6 +574,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -557,13 +586,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'Advertiser Id not found. Aborting',
+            error:
+              'Advertiser Id not found. Aborting: Workflow: procWorkflow, Step: validateInput, ChildStep: undefined, OriginalError: Advertiser Id not found. Aborting',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'configuration',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -645,6 +678,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -660,6 +694,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -672,13 +707,16 @@ export const data = [
         body: [
           {
             error:
-              'It is required at least one of em, hashed_maids or pair of client_ip_address and client_user_agent',
+              'It is required at least one of em, hashed_maids or pair of client_ip_address and client_user_agent: Workflow: procWorkflow, Step: validateUserFields, ChildStep: undefined, OriginalError: It is required at least one of em, hashed_maids or pair of client_ip_address and client_user_agent',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'instrumentation',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -763,6 +801,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -778,6 +817,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -789,6 +829,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -826,10 +869,16 @@ export const data = [
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
                       {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
                         quantity: 1,
                         item_price: '19',
                       },
                       {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
                         quantity: 2,
                         item_price: '3',
                       },
@@ -925,6 +974,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -940,6 +990,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -951,13 +1002,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'Advertiser Id not found. Aborting',
+            error:
+              'Advertiser Id not found. Aborting: Workflow: procWorkflow, Step: validateInput, ChildStep: undefined, OriginalError: Advertiser Id not found. Aborting',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'configuration',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -1041,6 +1096,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1056,6 +1112,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1067,13 +1124,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'message type group is not supported',
+            error:
+              'message type group is not supported: Workflow: procWorkflow, Step: validateInput, ChildStep: undefined, OriginalError: message type group is not supported',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'instrumentation',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -1158,6 +1219,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1173,6 +1235,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1184,13 +1247,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'Action source must be one of app_android, app_ios, web, offline',
+            error:
+              'Action source must be one of app_android, app_ios, web, offline: Workflow: procWorkflow, Step: validateCommonFields, ChildStep: undefined, OriginalError: Action source must be one of app_android, app_ios, web, offline',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'instrumentation',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -1275,6 +1342,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1291,6 +1359,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1302,6 +1371,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1339,10 +1411,16 @@ export const data = [
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
                       {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
                         quantity: 1,
                         item_price: '19',
                       },
                       {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
                         quantity: 2,
                         item_price: '3',
                       },
@@ -1402,6 +1480,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1417,6 +1496,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1428,6 +1508,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1512,6 +1595,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1527,6 +1611,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1538,6 +1623,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1622,6 +1710,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1637,6 +1726,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1648,6 +1738,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1733,6 +1826,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1748,6 +1842,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1759,6 +1854,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1845,6 +1943,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1858,6 +1957,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1869,6 +1969,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1957,6 +2060,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1970,6 +2074,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1981,6 +2086,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -2120,6 +2228,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2138,6 +2247,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2149,6 +2259,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               body: {
                 JSON: {
@@ -2240,6 +2353,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2256,6 +2370,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2267,6 +2382,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -2384,6 +2502,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2400,6 +2519,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2411,6 +2531,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -2446,7 +2569,15 @@ export const data = [
                     order_id: '50314b8e9bcf000000000000',
                     num_items: 1,
                     content_ids: ['507f1f77bcf86cd799439011'],
-                    contents: [{ quantity: 1, item_price: '19' }],
+                    contents: [
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                    ],
                   },
                 },
                 JSON_ARRAY: {},
@@ -2515,6 +2646,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2530,6 +2662,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2541,6 +2674,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               body: {
                 FORM: {},
@@ -2668,6 +2804,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2686,6 +2823,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2697,6 +2835,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -2740,10 +2881,16 @@ export const data = [
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
                       {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
                         quantity: 1,
                         item_price: '19',
                       },
                       {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
                         quantity: 2,
                         item_price: '3',
                       },
@@ -2840,6 +2987,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2858,6 +3006,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2869,6 +3018,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -2909,10 +3061,16 @@ export const data = [
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
                       {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
                         quantity: 1,
                         item_price: '19',
                       },
                       {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
                         quantity: 2,
                         item_price: '3',
                       },
@@ -3015,6 +3173,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -3033,6 +3192,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -3044,6 +3204,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -3108,10 +3271,16 @@ export const data = [
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
                       {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
                         quantity: 1,
                         item_price: '19',
                       },
                       {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
                         quantity: 2,
                         item_price: '3',
                       },
@@ -3171,6 +3340,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -3189,6 +3359,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -3200,13 +3371,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'Ad Account ID not found. Aborting',
+            error:
+              'Ad Account ID not found. Aborting: Workflow: procWorkflow, Step: validateInput, ChildStep: undefined, OriginalError: Ad Account ID not found. Aborting',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'configuration',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -3255,6 +3430,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -3273,6 +3449,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -3284,13 +3461,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'Conversion Token not found. Aborting',
+            error:
+              'Conversion Token not found. Aborting: Workflow: procWorkflow, Step: validateInput, ChildStep: undefined, OriginalError: Conversion Token not found. Aborting',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'configuration',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -3352,6 +3533,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -3370,6 +3552,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -3381,6 +3564,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
